refactor config file
