--- reddit/CONFIG.py
@@ -1,9 +1,0 @@
-data_path = "./data.json"
-keywords_path = "./keywords.json"
-creds_path = "./credentials.json"
-
-redis_creds = {
-    "host": 'redis-10381.c135.eu-central-1-1.ec2.cloud.redislabs.com',
-    "port": 10381,
-    "password": 'EcqJqZRW8CPeSWaSvBpmN4vOpIh2tXoZ'
-}
