+++ reddit/app.py
@@ -1,7 +1,7 @@
 from time import sleep
 from source.crawlers.reddit import RedditCrawler
-from source.utils.config import get_access, read, save, read_list
-from CONFIG import data_path, redis_creds
+from source.utils.tools import get_access_token, save_access_token, read_list
+import config
 import json
 import redis
 import logging
@@ -12,8 +12,8 @@ logging.basicConfig(
 
 logger = logging.getLogger(__name__)
 
-cache_db = redis.Redis(host=redis_creds["host"],
-                       port=redis_creds["port"], db=0, password=redis_creds["password"])
+cache_db = redis.Redis(host=config.redis_creds["host"],
+                       port=config.redis_creds["port"], db=0, password=config.redis_creds["password"])
 
 try:
     crawler = RedditCrawler(cache_db)
@@ -22,11 +22,10 @@ try:
         f'Logged in under account: {id["subreddit"]["title"]}|{id["subreddit"]["display_name_prefixed"]}')
 except:
     logger.info('Account is not logged in ! Initiate login ...')
-    creds = read()
-    response = get_access(
-        creds['USER_NAME'], creds['PASSWORD'], creds['CLIENT_ID'], creds['CLIENT_SECRET'])
-    creds['ACCESS_TOKEN'] = response['access_token']
-    save(creds)
+    response = get_access_token(
+        config.USER_NAME, config.PASSWORD, config.CLIENT_ID, config.CLIENT_SECRET)
+    access_token = response['access_token']
+    save_access_token(access_token)
     crawler = RedditCrawler(cache_db)
     id = crawler.identify()
     logger.info(
@@ -40,7 +39,7 @@ while True:
     for keyword in keywords:
         result[keyword] = crawler.hot_posts(keyword)
 
-    with open(data_path, 'w') as f:
+    with open(config.data_path, 'w') as f:
         json.dump(result, f)
 
     sleep(1)
