+++ reddit/config.py.template
@@ -1,0 +1,14 @@
+data_path = "./data.json"
+keywords_path = "./keywords.json"
+token_path = "./access_token.json"
+CLIENT_ID = ""
+CLIENT_SECRET = ""
+REDIRECT_URI =  ""
+USER_NAME = ""
+PASSWORD = ""
+
+redis_creds = {
+    "host": 'localhost',
+    "port": 6379,
+    "password": ''
+}
