--- reddit/credentials.json
@@ -1,8 +1,0 @@
-{
-    "CLIENT_ID": "pufiwxDrTOppPyZpG2QlJw",
-    "CLIENT_SECRET": "YMbz4g35Fvb-hrwj4iyj-tX-bLFNlA",
-    "REDIRECT_URI": "http://localhost:3000",
-    "USER_NAME": "Sad_Hall_7064",
-    "PASSWORD": "Hoangndnd1805!!!",
-    "ACCESS_TOKEN": "1832824652001-mol5zIxQewXWDOs7ZisS33NLK4WpZA"
-}
