+++ reddit/source/utils/tools.py
@@ -1,25 +1,25 @@
 import json
 import requests
 import requests.auth
-from CONFIG import creds_path, keywords_path
+import config
 
 def read_list():
-    with open(keywords_path, 'r') as f:
+    with open(config.keywords_path, 'r') as f:
         data = json.load(f)
         return data
 
-def read():
-    with open(creds_path, "r") as f:
+def read_access_token():
+    with open(config.token_path, "r") as f:
         data = json.load(f)
         return data
 
 
-def save(creds):
-    with open(creds_path, "w") as f:
+def save_access_token(creds):
+    with open(config.token_path, "w") as f:
         json.dump(creds, f)
 
 
-def get_access(username, password, client_id, client_secret):
+def get_access_token(username, password, client_id, client_secret):
     url = "https://www.reddit.com/api/v1/access_token"
     client_auth = requests.auth.HTTPBasicAuth(client_id, client_secret)
     post_data = {
@@ -35,7 +35,6 @@ def get_access(username, password, client_id, client_secret):
 
 
 if __name__ == "__main__":
-    creds = read()
-    access = get_access(creds["USER_NAME"], creds["PASSWORD"],
-                        creds["CLIENT_ID"], creds["CLIENT_SECRET"])
+    access = get_access_token(config.USER_NAME, config.PASSWORD,
+                        config.CLIENT_ID, config.CLIENT_SECRET)
     print(access)
